Refactoring status notification for user file upload, user fetch file, and user file deletion and correcting bug for deleteDocuments function
